network/requirements.txt
cryptography==2.7
configparser==3.7.4
Cython==0.29.12
ecos==2.0.7.post1
Flask==1.1.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gevent==1.4.0
glmnet-py==0.1.0b2
joblib==0.14.1
jsonschema==3.0.1
keyring==18.0.0
matplotlib==2.2.3
mkl-fft
mkl-random==1.0.1
multiprocess==0.70.10
networkx==2.2
numpy==1.16.4
numexpr==2.6.2
pandas==0.24.2
requests==2.22.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
tables
tqdm==4.32.1
cvxopt==1.2.5
scs==2.1.4
cvxpy==0.4.0
snap-stanford===5.0.0
